+++ src/opentimelineio/subtitles.h
@@ -22,46 +22,46 @@ namespace opentimelineio {
 
             using Parent = Item;
 
-            Subtitles(float extent_x = 0.f,
-                      float extent_y = 0.f,
-                      float padding_x = 0.f,
-                      float padding_y = 0.f,
+            Subtitles(double extent_x = 0.f,
+                      double extent_y = 0.f,
+                      double padding_x = 0.f,
+                      double padding_y = 0.f,
                       std::string background_color = std::string(),
-                      float background_opacity = 0.f,
+                      double background_opacity = 0.f,
                       DisplayAlignment display_alignment = DisplayAlignment::after,
                       std::vector<TimedText *> timed_texts = std::vector<TimedText *>());
 
             virtual TimeRange available_range(ErrorStatus *error_status) const;
 
-            float const extent_x() const {
+            double const extent_x() const {
                 return _extent_x;
             }
 
-            void set_extent_x(float const extent_x) {
+            void set_extent_x(double const extent_x) {
                 _extent_x = extent_x;
             }
 
-            float const extent_y() const {
+            double const extent_y() const {
                 return _extent_y;
             }
 
-            void set_extent_y(float const extent_y) {
+            void set_extent_y(double const extent_y) {
                 _extent_y = extent_y;
             }
 
-            float const padding_x() const {
+            double const padding_x() const {
                 return _padding_x;
             }
 
-            void set_padding_x(float const padding_x) {
+            void set_padding_x(double const padding_x) {
                 _padding_x = padding_x;
             }
 
-            float const padding_y() const {
+            double const padding_y() const {
                 return _padding_y;
             }
 
-            void set_padding_y(float const padding_y) {
+            void set_padding_y(double const padding_y) {
                 _padding_y = padding_y;
             }
 
@@ -73,11 +73,11 @@ namespace opentimelineio {
                 _background_color = background_color;
             }
 
-            float const background_opacity() const {
+            double const background_opacity() const {
                 return _background_opacity;
             }
 
-            void set_background_opacity(float const background_opacity) {
+            void set_background_opacity(double const background_opacity) {
                 _background_opacity = background_opacity;
             }
 
@@ -105,12 +105,12 @@ namespace opentimelineio {
             virtual void write_to(Writer &) const;
 
         private:
-            float _extent_x;
-            float _extent_y;
-            float _padding_x;
-            float _padding_y;
+            double _extent_x;
+            double _extent_y;
+            double _padding_x;
+            double _padding_y;
             std::string _background_color;
-            float _background_opacity;
+            double _background_opacity;
             DisplayAlignment _display_alignment;
             std::vector<Retainer < TimedText>> _timed_texts;
         };
